test(core): split pixiv.e2e.test.ts into per-resource e2e files
