--- packages/core/tests/e2e/pixiv.e2e.test.ts
@@ -1,571 +1,0 @@
-/**
- * End-to-end tests for PixivClient against the real pixiv API.
- *
- * These tests require a valid PIXIV_REFRESH_TOKEN environment variable.
- * They are skipped automatically when the token is not present, so they
- * never block CI runs that do not have the secret configured.
- *
- * Run manually:
- *   PIXIV_REFRESH_TOKEN=<token> pnpm --filter @book000/pixivts run test:e2e
- */
-
-import fs from 'node:fs'
-import { beforeAll, describe, expect, it } from 'vitest'
-import { PixivClient } from '../../src/client'
-
-// ---------------------------------------------------------------------------
-// Environment / skip guard
-// ---------------------------------------------------------------------------
-
-/** Load token from .env file if present, then from environment. */
-function loadToken(): string | undefined {
-  if (fs.existsSync('.env')) {
-    for (const line of fs.readFileSync('.env', 'utf8').split('\n')) {
-      const eq = line.indexOf('=')
-      if (eq === -1) continue
-      const key = line.slice(0, eq).trim()
-      const value = line.slice(eq + 1).trim()
-      if (key === 'PIXIV_REFRESH_TOKEN' && value) return value
-    }
-  }
-  return process.env.PIXIV_REFRESH_TOKEN
-}
-
-const REFRESH_TOKEN = loadToken()
-const SKIP = !REFRESH_TOKEN
-
-// ---------------------------------------------------------------------------
-// Test constants (stable public content)
-// ---------------------------------------------------------------------------
-
-/** A published illust (type: illust) used for stable assertions. */
-const ILLUST_ID = 107_565_629
-/** A published illust (type: manga) used for stable assertions. */
-const MANGA_ID = 103_905_962
-/** A published ugoira used for stable assertions. */
-const UGOIRA_ID = 83_638_393
-/** A published novel used for stable assertions. */
-const NOVEL_ID = 13_574_875
-/** An illust series used for stable assertions. */
-const ILLUST_SERIES_ID = 147_483
-/**
- * A novel series used for stable assertions.
- * Title: FGO夢まとめ (series ID 1458483, 58+ works, active since 2018)
- */
-const NOVEL_SERIES_ID = 1_458_483
-/** pixiv staff account (ID: 11) — always exists, safe for follow tests. */
-const STAFF_USER_ID = 11
-
-// ---------------------------------------------------------------------------
-// Suite
-// ---------------------------------------------------------------------------
-
-describe.skipIf(SKIP)('PixivClient e2e', () => {
-  let client: PixivClient
-
-  beforeAll(async () => {
-    // REFRESH_TOKEN is guaranteed non-empty here because of .skipIf(SKIP) above
-    client = await PixivClient.of(REFRESH_TOKEN ?? '')
-  })
-
-  // -------------------------------------------------------------------------
-  // Illusts
-  // -------------------------------------------------------------------------
-
-  it('illusts.detail — illust type', async () => {
-    const result = await client.illusts.detail({ illustId: ILLUST_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    const { illust } = result.value
-    expect(illust.id).toBe(ILLUST_ID)
-    expect(illust.type).toBe('illust')
-    expect(illust.user.id).toBe(16_668_308)
-    expect(illust.pageCount).toBe(1)
-    expect(illust.tags.length).toBeGreaterThan(0)
-    expect(illust.imageUrls.squareMedium).toMatch(
-      /^https:\/\/i\.pximg\.net\/.+\.jpg$/
-    )
-    expect(illust.totalBookmarks).toBeGreaterThan(0)
-    expect(illust.totalView).toBeGreaterThan(0)
-  })
-
-  it('illusts.detail — manga type', async () => {
-    const result = await client.illusts.detail({ illustId: MANGA_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    const { illust } = result.value
-    expect(illust.id).toBe(MANGA_ID)
-    expect(illust.type).toBe('manga')
-    expect(illust.pageCount).toBeGreaterThan(1)
-    expect(illust.metaPages.length).toBeGreaterThan(1)
-  })
-
-  it('illusts.related', async () => {
-    const result = await client.illusts.related({ illustId: ILLUST_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.related — with seedIllustIds filter', async () => {
-    const result = await client.illusts.related({
-      illustId: ILLUST_ID,
-      seedIllustIds: [ILLUST_ID],
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.search', async () => {
-    const result = await client.illusts.search({ word: 'ホロライブ' })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.ranking', async () => {
-    const result = await client.illusts.ranking({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.recommended', async () => {
-    const result = await client.illusts.recommended({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.series', async () => {
-    const result = await client.illusts.series({
-      illustSeriesId: ILLUST_SERIES_ID,
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illustSeriesDetail.title.length).toBeGreaterThan(0)
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  it('illusts.bookmarkAdd and illusts.bookmarkDelete', async () => {
-    // Check current state before mutating
-    const detailResult = await client.illusts.detail({ illustId: ILLUST_ID })
-    expect(detailResult.isOk).toBe(true)
-    if (!detailResult.isOk) return
-    const wasBookmarked = detailResult.value.illust.isBookmarked
-
-    try {
-      if (wasBookmarked) {
-        // delete then re-add to verify both operations work
-        const del = await client.illusts.bookmarkDelete({
-          illustId: ILLUST_ID,
-        })
-        expect(del.isOk).toBe(true)
-        const add = await client.illusts.bookmarkAdd({
-          illustId: ILLUST_ID,
-          restrict: 'public',
-          tags: ['テスト'],
-        })
-        expect(add.isOk).toBe(true)
-      } else {
-        // add then delete to verify both operations work
-        const add = await client.illusts.bookmarkAdd({
-          illustId: ILLUST_ID,
-          restrict: 'public',
-          tags: ['テスト'],
-        })
-        expect(add.isOk).toBe(true)
-        const del = await client.illusts.bookmarkDelete({
-          illustId: ILLUST_ID,
-        })
-        expect(del.isOk).toBe(true)
-      }
-    } finally {
-      // Restore original state
-      await (wasBookmarked
-        ? client.illusts.bookmarkAdd({
-            illustId: ILLUST_ID,
-            restrict: 'public',
-            tags: [],
-          })
-        : client.illusts.bookmarkDelete({ illustId: ILLUST_ID }))
-    }
-  })
-
-  it('illusts.follow', async () => {
-    const result = await client.illusts.follow({ restrict: 'public' })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.illusts)).toBe(true)
-  })
-
-  // Skipped: as of this writing, GET /v1/illust/comments returns an
-  // empty-bodied 404 against the live API for every path/param/method
-  // variant tried, while the same illust ID succeeds on /v1/illust/detail.
-  // This looks like a live-API-side issue (endpoint removed, or blocked by
-  // anti-bot filtering) rather than a bug in this method. Unskip once the
-  // live behavior is confirmed.
-  it.skip('illusts.comments', async () => {
-    const result = await client.illusts.comments({
-      illustId: ILLUST_ID,
-      includeTotalComments: true,
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.comments)).toBe(true)
-  })
-
-  it('illusts.bookmarkDetail', async () => {
-    const result = await client.illusts.bookmarkDetail({
-      illustId: ILLUST_ID,
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(typeof result.value.bookmarkDetail.isBookmarked).toBe('boolean')
-    expect(Array.isArray(result.value.bookmarkDetail.tags)).toBe(true)
-  })
-
-  it('illusts.new', async () => {
-    const result = await client.illusts.new({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  // -------------------------------------------------------------------------
-  // Manga
-  // -------------------------------------------------------------------------
-
-  it('manga.recommended', async () => {
-    const result = await client.manga.recommended({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.illusts.length).toBeGreaterThan(0)
-  })
-
-  // -------------------------------------------------------------------------
-  // Ugoira
-  // -------------------------------------------------------------------------
-
-  it('ugoira.metadata', async () => {
-    const result = await client.ugoira.metadata({ illustId: UGOIRA_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.ugoiraMetadata.zipUrls.medium).toMatch(
-      /^https:\/\/i\.pximg\.net\/img-zip-ugoira\/img\/.+_ugoira600x600\.zip$/
-    )
-    expect(result.value.ugoiraMetadata.frames.length).toBeGreaterThan(0)
-  })
-
-  // -------------------------------------------------------------------------
-  // Novels
-  // -------------------------------------------------------------------------
-
-  it('novels.detail', async () => {
-    const result = await client.novels.detail({ novelId: NOVEL_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novel.id).toBe(NOVEL_ID)
-    expect(result.value.novel.title.length).toBeGreaterThan(0)
-    expect(result.value.novel.pageCount).toBeGreaterThan(0)
-  })
-
-  it('novels.text', async () => {
-    const result = await client.novels.text({ novelId: NOVEL_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    // novels.text returns the raw novel text as a string
-    expect(result.value.length).toBeGreaterThan(0)
-  })
-
-  it('novels.related', async () => {
-    const result = await client.novels.related({ novelId: NOVEL_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.ranking', async () => {
-    const result = await client.novels.ranking({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.search', async () => {
-    const result = await client.novels.search({ word: 'ホロライブ' })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.recommended', async () => {
-    const result = await client.novels.recommended({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.series', async () => {
-    const result = await client.novels.series({ seriesId: NOVEL_SERIES_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novelSeriesDetail.id).toBe(NOVEL_SERIES_ID)
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.bookmarkAdd and novels.bookmarkDelete', async () => {
-    const detailResult = await client.novels.detail({ novelId: NOVEL_ID })
-    expect(detailResult.isOk).toBe(true)
-    if (!detailResult.isOk) return
-    const wasBookmarked = detailResult.value.novel.isBookmarked
-
-    try {
-      if (wasBookmarked) {
-        const del = await client.novels.bookmarkDelete({ novelId: NOVEL_ID })
-        expect(del.isOk).toBe(true)
-        const add = await client.novels.bookmarkAdd({
-          novelId: NOVEL_ID,
-          restrict: 'public',
-          tags: ['テスト'],
-        })
-        expect(add.isOk).toBe(true)
-      } else {
-        const add = await client.novels.bookmarkAdd({
-          novelId: NOVEL_ID,
-          restrict: 'public',
-          tags: ['テスト'],
-        })
-        expect(add.isOk).toBe(true)
-        const del = await client.novels.bookmarkDelete({ novelId: NOVEL_ID })
-        expect(del.isOk).toBe(true)
-      }
-    } finally {
-      await (wasBookmarked
-        ? client.novels.bookmarkAdd({
-            novelId: NOVEL_ID,
-            restrict: 'public',
-            tags: [],
-          })
-        : client.novels.bookmarkDelete({ novelId: NOVEL_ID }))
-    }
-  })
-
-  it('novels.follow', async () => {
-    const result = await client.novels.follow({ restrict: 'public' })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.novels)).toBe(true)
-  })
-
-  it('novels.follow with restrict=private', async () => {
-    const result = await client.novels.follow({ restrict: 'private' })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.novels)).toBe(true)
-  })
-
-  it('novels.follow with offset', async () => {
-    const result = await client.novels.follow({ restrict: 'public', offset: 1 })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.novels)).toBe(true)
-  })
-
-  it('novels.comments', async () => {
-    const result = await client.novels.comments({
-      novelId: NOVEL_ID,
-      includeTotalComments: true,
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.comments)).toBe(true)
-    expect(typeof result.value.totalComments).toBe('number')
-    if (result.value.commentAccessControl !== undefined) {
-      expect(typeof result.value.commentAccessControl).toBe('number')
-    }
-    if (result.value.comments.length > 0) {
-      const [comment] = result.value.comments
-      expect(typeof comment.id).toBe('number')
-      expect(typeof comment.comment).toBe('string')
-      expect(typeof comment.date).toBe('string')
-      expect(typeof comment.user.id).toBe('number')
-    }
-  })
-
-  it('novels.comments with offset', async () => {
-    const result = await client.novels.comments({
-      novelId: NOVEL_ID,
-      offset: 1,
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.comments)).toBe(true)
-  })
-
-  it('novels.new', async () => {
-    const result = await client.novels.new({})
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.novels.length).toBeGreaterThan(0)
-  })
-
-  it('novels.new resumes pagination via the maxNovelId cursor', async () => {
-    const first = await client.novels.new({})
-    expect(first.isOk).toBe(true)
-    if (!first.isOk) return
-    expect(first.value.novels.length).toBeGreaterThan(0)
-    if (first.value.nextUrl === null) return // no further pages to resume
-
-    // `next_url` embeds `max_novel_id`, but `ParsedNextUrl` does not expose
-    // it as a typed field — extract it manually from the raw URL.
-    const maxNovelId = Number(
-      new URL(first.value.nextUrl).searchParams.get('max_novel_id')
-    )
-    expect(Number.isNaN(maxNovelId)).toBe(false)
-
-    const second = await client.novels.new({ maxNovelId })
-    expect(second.isOk).toBe(true)
-    if (!second.isOk) return
-    expect(second.value.novels.length).toBeGreaterThan(0)
-  })
-
-  // -------------------------------------------------------------------------
-  // Users
-  // -------------------------------------------------------------------------
-
-  it('users.detail', async () => {
-    const result = await client.users.detail({ userId: STAFF_USER_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(result.value.user.id).toBe(STAFF_USER_ID)
-    expect(result.value.user.name.length).toBeGreaterThan(0)
-    expect(result.value.profile).toBeDefined()
-  })
-
-  it('users.illusts', async () => {
-    const result = await client.users.illusts({ userId: STAFF_USER_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    // The pixiv staff account may have no illusts; just check the shape.
-    expect(Array.isArray(result.value.illusts)).toBe(true)
-  })
-
-  it('users.novels', async () => {
-    const result = await client.users.novels({ userId: STAFF_USER_ID })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.novels)).toBe(true)
-  })
-
-  it('users.following', async () => {
-    const result = await client.users.following({
-      userId: STAFF_USER_ID,
-      restrict: 'public',
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.userPreviews)).toBe(true)
-  })
-
-  it('users.bookmarks.illusts', async () => {
-    const result = await client.users.bookmarks.illusts({
-      userId: STAFF_USER_ID,
-      restrict: 'public',
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.illusts)).toBe(true)
-  })
-
-  it('users.bookmarks.novels', async () => {
-    const result = await client.users.bookmarks.novels({
-      userId: STAFF_USER_ID,
-      restrict: 'public',
-    })
-    expect(result.isOk).toBe(true)
-    if (!result.isOk) return
-    expect(Array.isArray(result.value.novels)).toBe(true)
-  })
-
-  it('users.followAdd and users.followDelete', async () => {
-    const detailResult = await client.users.detail({ userId: STAFF_USER_ID })
-    expect(detailResult.isOk).toBe(true)
-    if (!detailResult.isOk) return
-    const wasFollowed = detailResult.value.user.isFollowed ?? false
-
-    try {
-      if (wasFollowed) {
-        const del = await client.users.followDelete({ userId: STAFF_USER_ID })
-        expect(del.isOk).toBe(true)
-        const add = await client.users.followAdd({
-          userId: STAFF_USER_ID,
-          restrict: 'public',
-        })
-        expect(add.isOk).toBe(true)
-      } else {
-        const add = await client.users.followAdd({
-          userId: STAFF_USER_ID,
-          restrict: 'public',
-        })
-        expect(add.isOk).toBe(true)
-        const del = await client.users.followDelete({ userId: STAFF_USER_ID })
-        expect(del.isOk).toBe(true)
-      }
-    } finally {
-      await (wasFollowed
-        ? client.users.followAdd({
-            userId: STAFF_USER_ID,
-            restrict: 'public',
-          })
-        : client.users.followDelete({ userId: STAFF_USER_ID }))
-    }
-  })
-
-  // -------------------------------------------------------------------------
-  // Images
-  // -------------------------------------------------------------------------
-
-  it('images.fetch', async () => {
-    // Fetch a known user avatar (small CDN image).
-    const detailResult = await client.users.detail({ userId: STAFF_USER_ID })
-    expect(detailResult.isOk).toBe(true)
-    if (!detailResult.isOk) return
-    const avatarUrl = detailResult.value.user.profileImageUrls.medium
-
-    // images.fetch returns ResultAsync<Response, PixivError>
-    const imgResult = await client.images.fetch(avatarUrl)
-    expect(imgResult.isOk).toBe(true)
-    if (!imgResult.isOk) return
-    expect(imgResult.value.ok).toBe(true)
-    expect(imgResult.value.status).toBe(200)
-  })
-
-  // -------------------------------------------------------------------------
-  // Pagination
-  // -------------------------------------------------------------------------
-
-  it('PaginatedResultAsync.pages() — iterates at least one page', async () => {
-    let pageCount = 0
-    let totalIllusts = 0
-    const pageIterable = client.illusts.search({ word: 'ホロライブ' }).pages()
-    for await (const page of pageIterable) {
-      pageCount++
-      totalIllusts += page.illusts.length
-      if (pageCount >= 2) break
-    }
-    expect(pageCount).toBeGreaterThanOrEqual(1)
-    expect(totalIllusts).toBeGreaterThan(0)
-  })
-
-  it('PaginatedResultAsync.items() — yields individual items', async () => {
-    const illusts: unknown[] = []
-    const itemIterable = client.illusts.search({ word: 'ホロライブ' }).items()
-    for await (const illust of itemIterable) {
-      illusts.push(illust)
-      if (illusts.length >= 60) break // stop after ~2 pages
-    }
-    expect(illusts.length).toBeGreaterThan(0)
-  })
-})
